feat: 调整变量名&修改模块结构
