+++ src/bridge/pojo/po/mod.rs
@@ -1,3 +1,5 @@
 pub mod bridge_user_ref_po;
+pub mod bridge_message_po;
 
 pub use bridge_user_ref_po::*;
+pub use bridge_message_po::*;
